requirements.test.txt
pytest>=7.0.0
pytest-homeassistant-custom-component>=0.10.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0

